requirements.txt
# Core RAG stack
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12

# PDF processing
PyMuPDF==1.24.14
camelot-py==1.0.0
opencv-python-headless==4.10.0.84
ghostscript==0.7
# pdfplumber==0.11.4

# Embeddings & retrieval
sentence-transformers==3.3.1
faiss-cpu==1.9.0.post1
rank-bm25==0.2.2

# LLM
google-generativeai==0.8.3

# Utilities
python-dotenv==1.0.1
requests==2.32.3
pydantic==2.10.3
